--- a/components/homme/src/share/cxx/HommexxEnums.hpp
+++ b/components/homme/src/share/cxx/HommexxEnums.hpp
@@ -9,6 +9,8 @@
 
 #include "Kokkos_Core.hpp"
 
+#include <cstdint>
+
 namespace Homme
 {
 
@@ -147,7 +149,7 @@ enum class DSSOption {
 // =================== Mesh connectivity enums ====================== //
 
 // The kind of connection: edge, corner or missing (one of the corner connections on one of the 8 cube vertices)
-enum class ConnectionKind : int {
+enum class ConnectionKind : std::uint8_t {
   EDGE    = 0,
   CORNER  = 1,
   MISSING = 2,  // Used to detect missing connections
@@ -155,14 +157,14 @@ enum class ConnectionKind : int {
 };
 
 // The locality of connection: local, shared or missing
-enum class ConnectionSharing : int {
+enum class ConnectionSharing : std::uint8_t {
   LOCAL   = 0,
   SHARED  = 1,
   MISSING = 2,  // Used to detect missing connections
   ANY     = 3   // Used when the kind of connection is not needed
 };
 
-enum class ConnectionName : int {
+enum class ConnectionName : std::uint8_t {
   // Edges
   SOUTH = 0,
   NORTH = 1,
@@ -178,7 +180,7 @@ enum class ConnectionName : int {
 
 // Direction (useful only for an edge)
 constexpr int NUM_DIRECTIONS = 3;
-enum class Direction : int {
+enum class Direction : std::uint8_t {
   FORWARD  = 0,
   BACKWARD = 1,
   INVALID  = 2