--- a/components/homme/src/share/cxx/mpi/BoundaryExchange.cpp
+++ b/components/homme/src/share/cxx/mpi/BoundaryExchange.cpp
@@ -8,7 +8,6 @@
 
 #include "MpiBuffersManager.hpp"
 #include "KernelVariables.hpp"
-#include "ErrorDefs.hpp"
 #include "profiling.hpp"
 
 #include "utilities/VectorUtils.hpp"