--- a/components/homme/src/theta-l_kokkos/cxx/HyperviscosityFunctorImpl.cpp
+++ b/components/homme/src/theta-l_kokkos/cxx/HyperviscosityFunctorImpl.cpp
@@ -17,8 +17,6 @@
 namespace Homme
 {
 
-#pragma message "TODO modify nutop policies according to nlev_tom"
-
 HyperviscosityFunctorImpl::
 HyperviscosityFunctorImpl (const SimulationParams&     params,
                            const ElementsGeometry&     geometry,