--- a/components/homme/src/theta-l_kokkos/cxx/HyperviscosityFunctorImpl.cpp
+++ b/components/homme/src/theta-l_kokkos/cxx/HyperviscosityFunctorImpl.cpp
@@ -202,6 +202,7 @@ void HyperviscosityFunctorImpl::init_boundary_exchanges () {
   std::shared_ptr<BoundaryExchange> bes[] = {m_be, m_be_tom};
   const int nlevs[] = {NUM_LEV, m_nu_scale_top_ilev_pack_lim};
   for (int i = 0; i < 2; ++i) {
+    if (i == 1 && m_data.nu_top <= 0) continue;
     auto be = bes[i];
     const auto nlev = nlevs[i];
     be->set_buffers_manager(bm_exchange);