… components/homme/src/share/cxx/SphereOperators.hpp
… components/homme/src/share/cxx/mpi/ConnectivityHelpers.hpp
… components/homme/src/share/cxx/mpi/mpi_cxx_f90_interface.cpp
… components/homme/src/share/cxx/prim_driver.cpp
… components/homme/src/share/cxx/prim_step.cpp
… components/homme/src/share/cxx/utilities/SyncUtils.hpp
… components/homme/src/share/gllfvremap_mod.F90
… components/homme/src/test_mod.F90
… components/homme/src/theta-l_kokkos/cxx/HyperviscosityFunctorImpl.hpp
… components/homme/src/theta-l_kokkos/cxx/cxx_f90_interface_theta.cpp
… components/homme/src/theta-l_kokkos/cxx/prim_advance_exp.cpp
… components/homme/src/theta-l_kokkos/element_state.F90
… components/homme/src/theta-l_kokkos/prim_driver_mod.F90
… components/homme/src/theta-l_kokkos/theta_f2c_mod.F90
… components/homme/test/reg_test/namelists/thetah-sl-test11conv-r0t1-cdr30-rrm.nl
… components/homme/test/reg_test/run_tests/test-list.cmake
… components/homme/test/reg_test/run_tests/thetah-sl-test11conv-r0t1-cdr30-rrm-cxx_vs_f90.cmake
… components/homme/test/reg_test/run_tests/thetah-sl-test11conv-r0t1-cdr30-rrm-kokkos.cmake
… components/homme/test_execs/share_kokkos_ut/geometry_interface.F90