* Rewrite Connectivity and BoundaryExchange impls to permit fully unstructured grids (RRM).
* Speed up top-of-model (TOM) dissipation in both kernels and communication.
* Add a new standalone-Homme C++-vs-F90 test for RRM support.

[BFB] except for a new RRM test in the HOMMEBFB test set.