@ambrad 2023-01-12
This PR requires checking the performance of two changes: the BE backend rewrite and the HV TOM improvements. The following performance results on Cori-KNL and Ascent show results for the old and new BE backends with the HV TOM improvements active, followed by the performance of master. The configuration is ne30pg2_ne30pg2.F2010-SCREAMv1.

Summary: The dycore gets a net small speedup on both platforms.

I need to run with two nodes instead of one on Cori-KNL, which increases variability substantially. The following timers are representative. In particular, both the new pack and unpack routines are faster than the old. The HV TOM improvement shows up in "a:tl-ae advance_hypervis_dp".
```
Cori-KNL (2 nodes x 64 ranks/node x 2 threads/rank; 1 node => OOM):
  BoundaryExchange backend:
    /global/cscratch1/sd/ambradl/e3sm_scratch/cori-knl/ne30pg2v1.nlev128
    old (timing.65819898.230110-150603)
        "a:amb_pack_min_max"           128 6.144000e+03   4.422064e+01     0.374   0.302
        "a:amb_unpack_min_max"         128 6.144000e+03   7.143616e+00     0.070   0.043
        "a:amb_pack_scon"              128 3.870720e+05   5.445228e+02     4.867   3.500
        "a:amb_unpack"                 128 3.870720e+05   6.677891e+02     5.580   4.818
        "a:tl-sc prim_run_subcycle_c"  128 6.144000e+03   1.660401e+04   129.902 129.575
        "a:caar compute"               128 3.686400e+05   3.697852e+03    29.872  27.738
        "a:caar_bexchV"                128 1.843200e+05   1.220387e+03    10.419   6.928
        "a:tl-ae advance_hypervis_dp"  128 3.686400e+04   1.217262e+03    10.314   8.577
        "a:hvf-bhwk"                   128 3.686400e+04   6.354812e+02     5.277   4.556
        "a:hvf-bexch"                  128 1.105920e+05   5.580245e+02     5.423   3.142
        "a:compose_transport"          128 6.144000e+03   4.144163e+03    32.622  32.164
    new (timing.65819614.230110-144424)
        "a:amb_pack_min_max"           128 6.144000e+03   4.385175e+01     0.386   0.315
        "a:amb_unpack_min_max"         128 6.144000e+03   8.459248e+00     0.085   0.057
        "a:amb_pack_ucon"              128 3.870720e+05   5.051612e+02     4.218   3.302
        "a:amb_unpack"                 128 3.870720e+05   5.276053e+02     4.508   3.598
        "a:tl-sc prim_run_subcycle_c"  128 6.144000e+03   1.665683e+04   130.427 129.949
        "a:caar compute"               128 3.686400e+05   3.779447e+03    30.811  28.571
        "a:caar_bexchV"                128 1.843200e+05   1.152549e+03    10.088   7.158
        "a:tl-ae advance_hypervis_dp"  128 3.686400e+04   1.229182e+03    10.378   8.685
        "a:hvf-bhwk"                   128 3.686400e+04   6.470293e+02     5.306   4.813
        "a:hvf-bexch"                  128 1.105920e+05   5.611220e+02     5.246   3.180
        "a:compose_transport"          128 6.144000e+03   4.110529e+03    32.519  31.917
  HV TOM changes to SphereOperators and BE 3D un/pack and comm:
    /global/cscratch1/sd/ambradl/e3sm_scratch/cori-knl/ne30pg2v1.ref.nlev128 (timing.65819576.230110-144022)
        "a:tl-sc prim_run_subcycle_c"  128 6.144000e+03   1.729269e+04   135.314 134.928
        "a:caar compute"               128 3.686400e+05   3.827495e+03    31.213  28.729
        "a:caar_bexchV"                128 1.843200e+05   1.280674e+03    11.081   8.087
        "a:tl-ae advance_hypervis_dp"  128 3.686400e+04   1.611209e+03    13.246  11.771
        "a:hvf-bhwk"                   128 3.686400e+04   6.802983e+02     5.601   4.993
        "a:hvf-bexch"                  128 1.105920e+05   7.193735e+02     6.466   4.398
        "a:compose_transport"          128 6.144000e+03   4.157903e+03    32.789  32.175
```

I can run on one node of Ascent. The new pack kernel is about the same speed as the old, while the new unpack kernel is a little less than 6% slower because of needing to access extra information about the corners, making the dycore ~0.4% slower in the absence of internode comm and thus negligibly slower when internode comm is present. Moreover, the HV TOM mod is enough to make the dycore slightly faster than on master, even on one node where the reduced HV TOM comm volume doesn't really matter. Thus, we've decided to remove the old BE backend rather than switch between structured and unstructured based on grid. Our choice has the advantage of minimizing code complexity and maximizing testing power.
```
Ascent (1 node x 6 ranks)
  BoundaryExchange backend:
    /gpfs/wolf/cli115/proj-shared/ambradl/e3sm_scratch/ne30pg2_ne30pg2.F2010-SCREAMv1.nn1
    old (timing.180088.230110-213601)
        "CPL:RUN_LOOP"                   6 2.880000e+02   4.313144e+02    71.888  71.882
        "CPL:ATM_RUN"                    6 2.880000e+02   2.784286e+02    47.986  44.830
        "a:amb_pack_min_max"             6 2.880000e+02   7.582726e-02     0.013   0.013
        "a:amb_unpack_min_max"           6 2.880000e+02   9.342753e-02     0.016   0.015
        "a:amb_pack_scon"                6 1.814400e+04   4.799743e+00     0.802   0.798
        "a:amb_unpack"                   6 1.814400e+04   9.372829e+00     1.568   1.549
        "a:tl-sc prim_run_subcycle_c"    6 2.880000e+02   1.317958e+02    21.969  21.963
        "a:caar compute"                 6 1.728000e+04   2.815378e+01     4.694   4.689
        "a:caar_bexchV"                  6 8.640000e+03   2.566613e+01     4.309   4.236
        "a:tl-ae advance_hypervis_dp"    6 1.728000e+03   2.004985e+01     3.375   3.316
        "a:hvf-bhwk"                     6 1.728000e+03   9.735696e+00     1.632   1.615
        "a:hvf-bexch"                    6 5.184000e+03   1.213789e+01     2.056   1.998
        "a:compose_transport"            6 2.880000e+02   1.328769e+01     2.221   2.209
    new (timing.180091.230110-220234)
        "CPL:RUN_LOOP"                   6 2.880000e+02   4.311858e+02    71.869  71.854
        "CPL:ATM_RUN"                    6 2.880000e+02   2.790474e+02    48.101  44.890
        "a:amb_pack_min_max"             6 2.880000e+02   7.597120e-02     0.013   0.013
        "a:amb_unpack_min_max"           6 2.880000e+02   9.151070e-02     0.016   0.015
        "a:amb_pack_ucon"                6 1.814400e+04   4.752973e+00     0.795   0.789
        "a:amb_unpack"                   6 1.814400e+04   9.870425e+00     1.659   1.637
        "a:tl-sc prim_run_subcycle_c"    6 2.880000e+02   1.322407e+02    22.043  22.036
        "a:caar compute"                 6 1.728000e+04   2.817030e+01     4.697   4.693
        "a:caar_bexchV"                  6 8.640000e+03   2.602906e+01     4.357   4.306
        "a:tl-ae advance_hypervis_dp"    6 1.728000e+03   2.005726e+01     3.353   3.337
        "a:hvf-bhwk"                     6 1.728000e+03   9.807070e+00     1.641   1.630
        "a:hvf-bexch"                    6 5.184000e+03   1.214217e+01     2.033   2.019
        "a:compose_transport"            6 2.880000e+02   1.337046e+01     2.231   2.224
  HV TOM changes to SphereOperators and BE 3D un/pack and comm:
    /gpfs/wolf/cli115/proj-shared/ambradl/e3sm_scratch/ne30pg2_ne30pg2.F2010-SCREAMv1.ref.nn1
    (timing.180089.230110-214453)
        "CPL:RUN_LOOP"                   6 2.880000e+02   4.332079e+02    72.207  72.183
        "CPL:ATM_RUN"                    6 2.880000e+02   2.814685e+02    48.612  45.135
        "a:tl-sc prim_run_subcycle_c"    6 2.880000e+02   1.351861e+02    22.533  22.530
        "a:caar compute"                 6 1.728000e+04   2.817866e+01     4.703   4.687
        "a:caar_bexchV"                  6 8.640000e+03   2.530603e+01     4.240   4.195
        "a:tl-ae advance_hypervis_dp"    6 1.728000e+03   2.383466e+01     3.992   3.960
        "a:hvf-bhwk"                     6 1.728000e+03   9.708187e+00     1.626   1.614
        "a:hvf-bexch"                    6 5.184000e+03   1.371808e+01     2.306   2.275
        "a:compose_transport"            6 2.880000e+02   1.327464e+01     2.220   2.198
```